Switch GitHub Actions workflow to use Nix
This is partially implementing a comment[1] from @drupol on a pull
request switching to use Nix Flakes:

> I'm wondering if we could do that in Github actions as well :-)

Right now only the builds and the artifact upload work, since there is
still an issue to solve with the artifacts.

Looking at the builds from the master branch, the artifacts generated
during the normal builds are real PDF files with the right mime type.

However, this isn't the case here and I haven't yet figured out why this
is the case, since what we get instead are ZIP files.

I also removed the dependencies job, since all information we need for
builds and for generating artifacts/releases can be fetched via
nix-instantiate.

[1]: https://github.com/hmemcpy/milewski-ctfp-pdf/pull/253#issuecomment-687111430

Signed-off-by: aszlig <aszlig@nix.build>