flake.nix: Split apart packages for reader/print
The GitHub Actions workflow builds separate packages for reader and
print and since we're going to use the Nix Flake to generate the build
matrix, we need to split it up here as well.

Since the GitHub Actions workflow has bypassed the Makefile, we're going
to do so here as well. This results less code for the Nix derivation.

Signed-off-by: aszlig <aszlig@nix.build>