fix: Continuous condition is not generalization of products of simpler morphisms
In the following three statements:

A. n -> m are products of simpler morphisms of types n -> 1.
B. A function that returns a product is a product of functionss.
C. The hom-functor is continuous.

the order of generalization should be C > A > B.
However, the original text can be read as if A > C.